CCCP-227, refactor: move to one's place
