+++ relayer/relayer.py
@@ -12,8 +12,8 @@ from chainpy.logger import global_logger
 from rbclib.__init__ import __version__
 from rbclib.bifrostutils import fetch_round_info, fetch_latest_round, find_height_by_timestamp, fetch_relayer_index
 from rbclib.consts import BIFROST_VALIDATOR_HISTORY_LIMIT_BLOCKS, BOOTSTRAP_OFFSET_ROUNDS
-from rbclib.globalconfig import RelayerRole, relayer_config_global
 from rbclib.switchable_enum import chain_primitives
+from relayer.global_config import RelayerRole, relayer_config_global
 
 
 class Relayer(EventBridge):
